% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brownfields_sf}
\alias{brownfields_sf}
\title{Brownfields and national priority sites}
\format{
An sf data frame with 150 rows and 11 variables:
\describe{
\item{id}{Integer. An ID, identical to the object ID in the original dataset.}
\item{name}{Character. Site name listed in the MDE database.}
\item{address}{Character. Site address(es).}
\item{city}{Character. Town name.}
\item{is_ongoing_assess}{Logical, whether assessment of the site is listed as ongoing.}
\item{is_ongoing_remed}{Logical, whether remediation of the site is listed as ongoing.}
\item{is_archived}{Logical, whether the site is considered closed.}
\item{fy_open}{Numeric. Fiscal year cleanup process was opened.}
\item{fy_closed}{Numeric. Fiscal year cleanup process was closed, if applicable.}
\item{site_type}{Factor. The site type (brownfield, npl, or both).}
\item{geometry}{POINT. Location.}
}
}
\source{
Maryland Department of the Environment Land Restoration Program, available at \url{https://mdewin64.mde.state.md.us/LRP/index.html}
}
\usage{
brownfields_sf
}
\description{
A \code{sf} data frame of basic information on brownfields and national priority list (superfund) sites in Maryland. This is a subset of data from the Maryland Department of the Environment's (MDE) Land Restoration Program, filtered for sites that are listed as brownfields, NPL sites, or both.
}
\examples{
head(brownfields_sf)
}
\keyword{datasets}
