% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{2022-based CPI inflation adjustment factor}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{year}{Numeric. Year of data.}
\item{adj_factor22}{Numeric. Factor by which to adjust dollar amounts to get 2022-based dollars.}
}
}
\source{
Bureau of Labor Statistics Consumer Price Index for All Urban Consumers (CPI-U) series CUUR0000SA0. \url{https://www.bls.gov/cpi/}
}
\usage{
cpi
}
\description{
A table of inflation adjustment factors for the year 1990 to 2022. This can be used to adjust dollar values for inflation into 2022 dollars, based on annual consumer price index (CPI) values from the Bureau of Labor Statistics. Match your values to \code{cpi} by year, then divide your values by the adjustment factor. Because these are annual averages, it won't be appropriate for more granular time periods in the past few years. This adjusts to 2022 dollars to match the \code{spending} dataset.
}
\examples{
  head(cpi)

  # to use cpi, join it to your data and divide values by the inflation factor
  set.seed(1)
  x <- data.frame(year = 2016:2022, value = round(rnorm(7, 100, 10)))
  x |>
    dplyr::left_join(cpi, by = "year") |>
    dplyr::mutate(adj_value = value / adj_factor22)
}
\keyword{datasets}
