% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Local normalized diversity index}
\usage{
diversity(x)
}
\arguments{
\item{x}{A data frame containing only numeric columns, where each column represents counts of population groups}
}
\value{
A numeric vector of index values, with length the same as the number of rows in \code{x}
}
\description{
This is a rewrite of \code{OasisR::HLoc}, an implementation of the Shannon-Weiner entropy-based diversity index. It takes a data frame of population counts (not shares!), where each row is one location and each column is a mutually exclusive demographic group. It then returns the index for each location.
}
\examples{
  race <- data.frame(white = c(3394, 3036, 1527),
                     black = c(13, 779, 448),
                     latino = c(38, 99, 125),
                     asian = c(0, 18, 34),
                     other_race = c(29, 120, 170))
 diversity(race)
}
\seealso{
\link[OasisR:HLoc]{OasisR::HLoc}
}
