% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ej_natl}
\alias{ej_natl}
\title{EPA environmental justice index}
\format{
A data frame with 18879 rows and 5 variables:
\describe{
\item{tract}{Character. The tract FIPS code.}
\item{indicator}{Factor. The environmental health risk factor, such as proximity to water treatment or air pollution-related cancers.}
\item{value_ptile}{Integer. The nationwide percentile of indexed values.}
\item{d2_ptile}{Integer. The percentile of indexed values scaled based on a two-factor demographic index (percent low-income and percent people of color).}
\item{d5_ptile}{Integer. The percentile of indexed values scaled based on a five-factor demographic index (percent low-income, unemployment rate, percent limited English, percent without high school diploma, low life expectancy).}
}
}
\source{
Environmental Protection Agency (EPA) EJSCREEN Environment Justice Index. ~~Data portal, definitions, and methodology are available at \url{https://www.epa.gov/ejscreen/technical-information-about-ejscreen}~~ Removed in early 2025 from EPA servers by DOGE, but many people and organizations host backup copies. For this package, the data comes from an archive at Harvard Dataverse. EPA. (2024). Environmental justice mapping and screening tool (EJScreen) (Version 4.0) [Dataset]. Harvard Dataverse. https://doi.org/10.7910/DVN/RLR5AX
}
\usage{
ej_natl
}
\description{
A dataset containing environmental health risk factors from the EPA's EJSCREEN environment justice index for census tracts in Maryland. Values are calculated based on aggregations of risk factors, then given as percentiles compared to all tracts across the US. Columns starting with \code{"d"} are adjusted for one of two different definitions of vulnerable populations.
}
\examples{
head(ej_natl)
}
\seealso{
\href{https://dataverse.harvard.edu/file.xhtml?fileId=10775982&version=4.0}{EJSCREEN technical docs}
}
\keyword{datasets}
