% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ej_trend}
\alias{ej_trend}
\title{EPA environmental justice index trend}
\format{
A data frame with 270334 rows and 6 variables:
\describe{
\item{year}{Numeric. Year of data.}
\item{bg}{Character. The block group FIPS code.}
\item{total_pop}{Numeric. Total population of the block group that year.}
\item{indicator}{Factor. The environmental health risk factor, such as proximity to water treatment or air pollution-related cancers.}
\item{value_ptile}{Integer. The nationwide percentile of indexed values.}
\item{d2_ptile}{Integer. The percentile of indexed values scaled based on a two-factor demographic index (percent low-income and percent people of color).}
}
}
\source{
Environmental Protection Agency (EPA) EJSCREEN Environment Justice Index. Data portal, definitions, and methodology are available at \url{https://www.epa.gov/ejscreen/technical-information-about-ejscreen}
}
\usage{
ej_trend
}
\description{
A dataset containing environmental health risk factors from the EPA's EJSCREEN environment justice index for block groups in Maryland for each year from 2018 to 2023. Values are calculated based on aggregations of risk factors, then given as nationwide percentiles. The column \code{d2_ptile} is adjusted for the EPA's two-factor definition of vulneration populations. While different years of this data are formatted differently, all indicators subset here are consistent except for underground storage, which is only available starting in 2021. Note also that block group definitions and their GEOIDs changed with the 2020 decennial census: 2018 to 2021 use 2010 block groups, while 2022 and 2023 use 2020 block groups. If joining with shapefiles from TIGER or elsewhere, you'll need one for 2010 block groups and one for 2020.
}
\examples{
head(ej_trend)
}
\seealso{
ejscreen
}
\keyword{datasets}
