% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejscreen}
\alias{ejscreen}
\title{EPA environmental justice index}
\format{
A data frame with 19175 rows and 5 variables:
\describe{
\item{tract}{Character. The tract FIPS code.}
\item{indicator}{Factor. The environmental health risk factor, such as proximity to water treatment or air pollution-related cancers.}
\item{value_ptile}{Integer. The nationwide percentile of indexed values.}
\item{d2_ptile}{Integer. The percentile of indexed values scaled based on a two-factor demographic index (percent low-income and percent people of color).}
\item{d5_ptile}{Integer. The percentile of indexed values scaled based on a five-factor demographic index (percent low-income, unemployment rate, percent limited English, percent without high school diploma, low life expectancy).}
}
}
\source{
Environmental Protection Agency (EPA) EJSCREEN Environment Justice Index. Data portal, definitions, and methodology are available at \url{https://www.epa.gov/ejscreen/technical-information-about-ejscreen}
}
\usage{
ejscreen
}
\description{
A dataset containing environmental health risk factors from the EPA's EJSCREEN environment justice index for census tracts in Maryland. Values are calculated based on aggregations of risk factors, then given as nationwide percentiles. Columns starting with \code{"d"} are adjusted for one of two different definitions of vulnerable populations.
}
\keyword{datasets}
