% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_land_cover.R
\name{load_land_cover}
\alias{load_land_cover}
\title{Load land cover raster data}
\source{
\url{https://www.mrlc.gov/data/legends/national-land-cover-database-class-legend-and-description}
}
\usage{
load_land_cover(type = c("canopy", "impervious"))
}
\arguments{
\item{type}{String, type of land cover raster to return. Either "canopy" or "impervious".}
}
\value{
A raster of class \code{terra::SpatRaster} with 1 band.
}
\description{
This is a helper function to load one of two types of raster data for the Baltimore area. Both are land cover data from the National Land Cover Database (NLCD) obtained with ROpenSci's \code{FedData} package. The "canopy" data is the percentage of each pixel that is covered by tree canopy, for vegetation-related land cover types. The "impervious" data is the percentage of each pixel that is considered an impervious surface, for urban developed land cover types. Both files originally were downloaded as 30m x 30m rasters with values ranging from 1 to 255 and \code{NA} at pixels of other land cover types. Both were then downsampled to 90m x 90m, scaled to percentages, and \code{NA} values were replaced with 0.
}
\examples{
canopy <- load_land_cover("canopy")
canopy
\dontrun{
  if (interactive()) {
    plot(canopy)
  }
}
}
\seealso{
\code{\link[FedData:get_nlcd]{FedData::get_nlcd}}
}
