% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonres_sf}
\alias{nonres_sf}
\title{Shapefile of nonresidential areas}
\format{
An sf data frame with 514 rows and 5 variables:
\describe{
\item{type}{Factor. Site type: one of airport, industrial, military, prison, or protected area. These refer to the keys used to retrieve data from OSM, though some may have matched multiple keys (duplicates were removed).}
\item{osm_id}{Character. The OpenStreetMap ID for the site; can be uesd to retrieve more metadata.}
\item{name}{Character. The name of the site.}
\item{geometry}{POLYGON. The boundaries of the site.}
\item{is_low_res}{Logical. Gives whether the site intersects with a low-residential tract, defined as tracts with fewer than 500 households based on the 2022 ACS.}
}
}
\source{
OpenStreetMap database via the \href{https://github.com/ropensci/osmdata}{\code{osmdata}} package, and American Community Survey 2022 5-year estimates via \href{https://github.com/walkerke/tidycensus}{\code{tidycensus}}.
}
\usage{
nonres_sf
}
\description{
An \code{sf} object of data from OpenStreetMap of large nonresidential areas in Maryland. There are several census tracts with very few households that coincide with industrial sites, prisons, and other nonresidential properties, and this is an attempt to identify some of them.
}
\examples{
head(nonres_sf)
}
\keyword{datasets}
