% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_stops}
\alias{police_stops}
\title{Police stops in Connecticut}
\format{
A data frame with 13 rows and 5 variables:
\describe{
\item{name}{Character. The name of the location where the data was collected.}
\item{category}{Factor. Category for which data are aggregated (total, gender, or race/ethnicity).}
\item{group}{Factor. The group within the given category for which data are aggregated (Total, Male, Female, White, Black, Latino).}
\item{ever_unfairly_stopped}{Numeric. The share of respondents who reported having ever been unfairly stopped, harassed, or abused by police.}
\item{multiple_times_3yr}{Numeric. The share of respondents who reported this treatment having happened multiple times in the past 3 years.}
}
}
\source{
DataHaven Community Wellbeing Survey 2021, analyzed with Camille's packages \code{\link{cwi}} and \code{\link{dcws}}. \url{https://ctdatahaven.org/reports/datahaven-community-wellbeing-survey}
}
\usage{
police_stops
}
\description{
A dataset containing responses to questions about unfair treatment by police from the 2021 DataHaven Community Wellbeing Survey. The data is broken down by location, category, and group, with values for adults in Connecticut, New Haven, and the Greater New Haven area.
}
\keyword{datasets}
