% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sum100.R
\name{round_sum100}
\alias{round_sum100}
\title{Round a set of numbers so they'll add up to 100}
\usage{
round_sum100(x, digits = 0, verbose = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Number of digits to use for rounding; you probably don't need
to change this. Default: 0}

\item{verbose}{Logical; if \code{TRUE}, will print the sum, letting you confirm
whether the numbers do indeed add to 100. Defaults \code{FALSE}.}
}
\value{
A numeric vector that hopefully adds up to 100.
}
\description{
This function takes a vector of numbers and rounds them to
coerce them to add up to 100. It works best for numbers with several decimal
places that are close to adding to 100. It's definitely not
fool-proof, but can help with situations like waffle charts where previous
rounding may make numbers not add up to exactly 100.
}
\examples{
round_sum100(c(9.2124, 40.292, 50.2), 0) # yay
round_sum100(c(0.24, 0.61, 0.15) * 100) # yay
round_sum100(c(0.24, 0.6, 0.15) * 100) # sad
}
