% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracts_sf}
\alias{tracts_sf}
\title{Census tracts for the Baltimore area}
\format{
An sf data frame with 606 rows and 3 variables:
\describe{
\item{county}{Character. The name of the county where the census tract is located.}
\item{geoid}{Character. The FIPS code for the census tract.}
\item{geometry}{POLYGON. The geometric representation of the census tract boundary.}
}
}
\source{
U.S. Census Bureau, TIGER boundary files via the \href{https://github.com/walkerke/tigris/}{\code{tigris}} package
}
\usage{
tracts_sf
}
\description{
A dataset containing 2020 census tract boundaries for Baltimore and surrounding counties. The data comes from the \code{tigris} package with geometries simplified.
}
\examples{
 head(tracts_sf)
}
\keyword{ref-datasets}
