% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trashwheel}
\alias{trashwheel}
\title{Collection of trash by Mr. Trash Wheel and friends}
\format{
A data frame with 1313 rows and 12 variables:
\describe{
\item{name}{Factor. Name of trash wheel.}
\item{dumpster}{Numeric. Dumpster number within the trash wheel.}
\item{date}{Date of counting.}
\item{weight_tons}{Numeric. Total weight of trash collected in tons.}
\item{volume_cubic_yards}{Numeric. Total volume of trash collected in cubic yards.}
\item{plastic_bottles}{Numeric. Estimated number of plastic bottles collected.}
\item{polystyrene}{Numeric. Estimated number of pieces of polystyrene collected.}
\item{cigarette_butts}{Numeric. Estimated number of cigarette butts collected.}
\item{glass_bottles}{Numeric. Estimated number of glass bottles collected.}
\item{plastic_bags}{Numeric. Estimated number of plastic bags collected.}
\item{wrappers}{Numeric. Estimated number of wrappers collected.}
\item{sports_balls}{Numeric. Estimated number of sports balls collected.}
}
}
\source{
Waterfront Partnership of Baltimore. (2026). Trash Interception. Mr. Trash Wheel. https://www.mrtrashwheel.com/trash-interception
}
\usage{
trashwheel
}
\description{
Amounts of trash collected by each of the 4 trash wheels in Baltimore's harbor, marked by date. Includes total weight and volume, then estimates of counts of items by type. Values may be \code{NA} if counts of an item were not available for a trash wheel.
}
\examples{
  head(trashwheel)
}
\keyword{datasets}
