% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacant}
\alias{vacant}
\title{Vacancy rates and correlated housing measures}
\format{
A data frame with 264 rows and 11 variables:
\describe{
\item{state}{Character. The state FIPS code.}
\item{city}{Character. The city name.}
\item{county}{Character. The 3-digit county FIPS code.}
\item{geoid}{Character. The tract FIPS code.}
\item{total_units}{Numeric. The total number of housing units.}
\item{vacants}{Numeric. The number of vacant housing units.}
\item{med_rent}{Numeric. The median rent for renter-occupied housing units.}
\item{med_housing_value}{Numeric. The median housing value for owner-occupied housing units.}
\item{vacancy_rate}{Numeric. The vacancy rate.}
\item{area_sqmi}{Numeric. The area in square miles.}
\item{housing_density}{Numeric. The housing density (total units per square mile).}
}
}
\source{
Calculated from US Census Bureau. American Community Survey 2022 5-year estimates. Calculated by Camille with the \href{https://github.com/walkerke/tidycensus}{\code{tidycensus}} and \href{https://github.com/CT-Data-Haven/cwi}{\code{cwi}} packages.
}
\usage{
vacant
}
\description{
A dataset containing vacancy rates and other values related to housing for tracts in Baltimore; Stamford, CT; and New Haven, CT. The data comes from the 2022 American Community Survey (ACS).
}
\keyword{datasets}
