% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages}
\alias{wages}
\title{Median wages by demographic}
\format{
A data frame with 59 rows and 12 variables:
\describe{
\item{dimension}{Factor. The dimension across which values are calculated.}
\item{status}{Factor. Worker status: all workers, full-time workers, or part-time workers. Some groups are only available for full-time workers.}
\item{sex}{Factor. The sex of the individuals.}
\item{race_eth}{Factor. The race/ethnicity of the individuals.}
\item{edu}{Factor. The education level of the individuals.}
\item{count}{Numeric. The estimated number of individuals in the group.}
\item{sample_n}{Numeric. The sample size used for estimates.}
\item{earn_q20}{Numeric. The 20th percentile of earnings.}
\item{earn_q25}{Numeric. The 25th percentile of earnings.}
\item{earn_q50}{Numeric. The 50th percentile (median) earnings.}
\item{earn_q75}{Numeric. The 75th percentile of earnings.}
\item{earn_q80}{Numeric. The 80th percentile of earnings.}
}
}
\source{
U.S. Census Bureau, American Community Survey, Integrated Public Use Microdata Series \url{https://usa.ipums.org/usa/}. Analyzed using the \href{https://github.com/gergness/srvyr}{\code{srvyr}} package.
}
\usage{
wages
}
\description{
A dataset containing median individual earnings by various dimensions (sex, race, education, etc.) for Maryland, for adults ages 25 and up with positive earnings. The data is calculated from the 2024 American Community Survey (ACS) Public Use Microdata Sample (PUMS) data via the Integrated Public Use Microdata Series (IPUMS).
}
\examples{
 head(wages)
}
\keyword{ref-datasets}
