% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages_by_puma}
\alias{wages_by_puma}
\title{Median wages by PUMA}
\format{
A data frame with 4702 rows and 5 variables:
\describe{
\item{statefip}{Character. The state FIPS code.}
\item{geoid}{Character. The PUMA FIPS code.}
\item{sex}{Character. The sex of the individuals.}
\item{count}{Numeric. The number of individuals.}
\item{earn}{Numeric. The median earnings.}
}
}
\source{
U.S. Census Bureau, American Community Survey, Integrated Public Use Microdata Series \url{https://usa.ipums.org/usa/}. Analyzed using the \code{\link{srvyr}} package.
}
\usage{
wages_by_puma
}
\description{
A dataset containing median individual earnings by sex for all Public Use Microdata Areas (PUMAs) in the US, for adults ages 25 and up working full time with positive earnings. The data is calculated from the 2022 American Community Survey (ACS) Public Use Microdata Sample (PUMS) data via the Integrated Public Use Microdata Series (IPUMS). PUMAs are areas of at least 100,000 people.
}
\keyword{datasets}
