% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk_tract_10_to_20}
\alias{xwalk_tract_10_to_20}
\title{2010 to 2020 tract crosswalk}
\format{
A data frame with 2052 rows and 3 variables:
\describe{
\item{tract20}{Character. Tract GEOID based on 2020 geographies.}
\item{tract10}{Character. Tract GEOID based on 2010 geographies.}
\item{weight}{Numeric. Weight to use in converting values based on 2010 geographies into ones based on 2020 geographies. Use this as the weight for weighted mean calculations.}
}
}
\source{
Block-to-block crosswalk from IPUMS NHGIS, University of Minnesota, www.nhgis.org.
}
\usage{
xwalk_tract_10_to_20
}
\description{
The Census Bureau updates the boundaries of its geographies, such as tracts, after every decennial census. To convert values based on one year's geographies to another year's geographies, you need what's called a crosswalk, or a table of weights to use for averaging values. Unfortunately, many results of the 2020 census were delayed, so some datasets are still put out with 2010 geographies, while others have updated to 2020. The CDC Places data still uses 2010 boundaries, so if you want to merge data from \code{cdc} to data from any of the other tract-level datasets, you'll need to use this crosswalk. See the example below.
}
\examples{
head(xwalk_tract_10_to_20)

# Here's an example walkthrough of how you would prepare the CDC data to join
# it with another dataset---I'll use the EJSCREEN data. I'm filtering each of
# these for the indicators I'm interested in, which are asthma from CDC and
# traffic from EJSCREEN.

library(dplyr)

# filter cdc data for just asthma
asthma10 <- cdc |>
  filter(indicator == "Current asthma", level == "tract") |>
  select(tract10 = location, asthma = value, pop)

# calculate values based on 2020 geographies:
# - average asthma rate is calculated as a weighted mean of rates
# - average adult population is calculated as a weighted sum of counts
# (I'm not actually using adult pop for anything, just here as an example of weighting counts)
asthma20 <- asthma10 |>
  inner_join(xwalk_tract_10_to_20, by = "tract10") |>
  group_by(tract20) |>
  summarise(asthma = weighted.mean(asthma, weight),
            adult_pop = sum(pop * weight))

# filter ejscreen for just traffic, rename value column accordingly
traffic <- ejscreen |>
  filter(indicator == "traffic") |>
  select(tract20 = tract, traffic_value_ptile = value_ptile)

# join both data frames now that they have geographies in common
traffic |>
  inner_join(asthma20, by = "tract20")
}
\keyword{datasets}
