% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{youth_risks}
\alias{youth_risks}
\title{Predictions of youth risks in Connecticut}
\format{
A data frame with \code{nrow(youth_risks)} rows and \code{ncol(youth_risks)} variables:
\describe{
\item{code}{Character. The code for the question.}
\item{question}{Character. The question about the predicted outcome for youth.}
\item{category}{Factor. The category of the respondent (e.g., total, gender).}
\item{group}{Factor. The group of the respondent (e.g., Connecticut, male, female).}
\item{response}{Factor. The response to the question (almost certain, very likely, a toss up, not very likely, or not at all likely).}
\item{value}{Numeric. The share of respondents giving the response.}
}
}
\source{
DataHaven Community Wellbeing Survey 2021, analyzed with Camille's packages \href{https://github.com/CT-Data-Haven/cwi}{\code{cwi}} and \code{dcws} (not currently public). \url{https://ctdatahaven.org/reports/datahaven-community-wellbeing-survey}
}
\usage{
youth_risks
}
\description{
A dataset containing Likert-style responses to questions about different outcomes for youth in their area, from the 2021 DataHaven Community Wellbeing Survey. The data is broken down by category, group, and response with values for adults in Connecticut.
}
\keyword{datasets}
