% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs}
\alias{acs}
\title{Demographics and socio-economic indicators from the 2022 ACS}
\format{
A data frame with 1870 rows and 30 variables:
\describe{
\item{level}{Factor. Geographic level (us, msa, state, county, or tract).}
\item{county}{Character. Name of the county for tracts, \code{NA} otherwise.}
\item{name}{Character. The name of the geography, including FIPS codes for tracts.}
\item{total_pop}{Numeric. Total population.}
\item{white}{Numeric. Share of population that is White.}
\item{black}{Numeric. Share of population that is Black.}
\item{latino}{Numeric. Share of population that is Latino.}
\item{asian}{Numeric. Share of population that is Asian.}
\item{other_race}{Numeric. Share of population that is of other race.}
\item{diversity_idx}{Numeric. Diversity index based on the preceding race/ethnicity columns. Uses Theil's H entropy.}
\item{foreign_born}{Numeric. Share of population that is foreign born.}
\item{total_hh}{Numeric. Total households.}
\item{homeownership}{Numeric. Homeownership rate.}
\item{total_cost_burden}{Numeric. Share of households that are cost burdened, based on HUD's standard that housing costs should be no more than 30\% of a household's total income.}
\item{total_severe_cost_burden}{Numeric. Share of households that are severely cost burdened, or paying more than 50\% of their income toward housing costs.}
\item{owner_cost_burden}{Numeric. Share of homeowners that are cost burdened.}
\item{owner_severe_cost_burden}{Numeric. Share of homeowners that are severely cost burdened.}
\item{renter_cost_burden}{Numeric. Share of renters that are cost burdened.}
\item{renter_severe_cost_burden}{Numeric. Share of renters that are severely cost burdened}
\item{no_vehicle_hh}{Numeric. Share of households without a vehicle.}
\item{median_hh_income}{Numeric. Median household income in 2022 dollars.}
\item{ages25plus}{Numeric. Population aged 25 and over.}
\item{less_than_high_school}{Numeric. Share of population aged 25 and over with less than a high school diploma.}
\item{high_school_grad}{Numeric. Share of population aged 25 and over with a high school diploma.}
\item{some_college_or_aa}{Numeric. Share of population aged 25 and over with some college or an associate degree.}
\item{bachelors}{Numeric. Share of population aged 25 and over with a bachelor's degree.}
\item{grad_degree}{Numeric. Share of population aged 25 and over with a graduate degree.}
\item{pov_status_determined}{Numeric. Population for whom poverty status is determined.}
\item{poverty}{Numeric. Poverty rate, or the share of the population for whom poverty status is determined that lives in a household with income below the federal poverty level.}
\item{low_income}{Numeric. Low-income rate, or the share of the population for whom poverty status is determined that lives in a household with income below 2 times the federal poverty level.}
}
}
\source{
Calculated from US Census Bureau. American Community Survey 2022 5-year estimates. Calculated by Camille with the \href{https://github.com/walkerke/tidycensus}{\code{tidycensus}} and \href{https://github.com/CT-Data-Haven/cwi}{\code{cwi}} packages.
}
\usage{
acs
}
\description{
A dataset containing indicators from the US Census Bureau's 2022 American Community Survey 5-year estimates. These are given for several geographic levels, including the US, every metropolitan statistical area (MSA) in the country, the state of Maryland, and every county and census tract in Maryland.
}
\keyword{datasets}
