% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highways_sf}
\alias{highways_sf}
\title{Shapefile of highways}
\format{
An sf data frame with 8327 rows and 4 variables:
\describe{
\item{osm_id}{Character. The OpenStreetMap ID for the highway; can be used to retrieve more metadata.}
\item{name}{Character. The name of the highway, if labeled in the OSM database.}
\item{lanes}{Numeric. The number of lanes on the highway.}
\item{geometry}{LINESTRING. The geometric representation of the highway.}
}
}
\source{
OpenStreetMap database via the \href{https://github.com/ropensci/osmdata}{\code{osmdata}} package.
}
\usage{
highways_sf
}
\description{
An \code{sf} object containing highways (specifically keyed as motorways or trunks) in Baltimore city and surrounding counties from OpenStreetMap.
}
\examples{
 head(highways_sf)
}
\keyword{datasets}
