% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{Monthly unemployment rates}
\format{
A data frame with 7800 rows and 3 variables:
\describe{
\item{name}{Character. The name of the location.}
\item{date}{Date. The month for which unemployment is reported.}
\item{rate}{Numeric. The reported unemployment rate.}
}
}
\source{
U.S. Bureau of Labor Statistics, Local Area Unemployment Statistics via API with the \href{https://github.com/CT-Data-Haven/cwi}{\code{cwi}} package. \url{https://www.bls.gov/lau/}
}
\usage{
unemployment
}
\description{
A dataset containing monthly unemployment rates from 2000 to 2025 for Maryland, Baltimore city, and all counties in the state. The data comes from the Bureau of Labor Statistics' Local Area Unemployment Statistics (LAUS).
}
\examples{
 head(unemployment)
}
\keyword{datasets}
