% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{Monthly unemployment rates}
\format{
A data frame with 1436 rows and 4 variables:
\describe{
\item{name}{Character. The name of the location.}
\item{date}{Date. The month for which unemployment is reported.}
\item{reported_rate}{Numeric. The reported unemployment rate.}
\item{adjusted_rate}{Numeric. The seasonally adjusted unemployment rate.}
}
}
\source{
U.S. Bureau of Labor Statistics, Local Area Unemployment Statistics via API with the \href{https://github.com/CT-Data-Haven/cwi}{\code{cwi}} package. \url{https://www.bls.gov/lau/} Seasonal adjustment is done with the BLS' methodology via \code{\link[feasts:X_13ARIMA_SEATS]{feasts::X_13ARIMA_SEATS}}.
}
\usage{
unemployment
}
\description{
A dataset containing monthly unemployment rates from 2000 to 2023 for Maryland, Baltimore city, and surrounding counties. The data comes from the Bureau of Labor Statistics' Local Area Unemployment Statistics (LAUS).
}
\examples{
 head(unemployment)
}
\keyword{datasets}
