% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages}
\alias{wages}
\title{Median wages by demographic}
\format{
A data frame with 309 rows and 15 variables:
\describe{
\item{dimension}{Factor. The dimension across which values are calculated.}
\item{name}{Factor. The name of the region (US or Maryland).}
\item{sex}{Factor. The sex of the individuals.}
\item{race_eth}{Factor. The race/ethnicity of the individuals.}
\item{edu}{Factor. The education level of the individuals.}
\item{occ_group}{Factor. The occupation group of the individuals.}
\item{univ}{Factor. Universe of workers, whether the denominator is all workers or full-time only.}
\item{is_fulltime}{Logical. Whether the individual is a full-time worker. \code{NA} if distinction isn't included.}
\item{count}{Numeric. The number of individuals in the group.}
\item{earn_q20}{Numeric. The 20th percentile of earnings.}
\item{earn_q25}{Numeric. The 25th percentile of earnings.}
\item{earn_q50}{Numeric. The 50th percentile (median) earnings.}
\item{earn_q75}{Numeric. The 75th percentile of earnings.}
\item{earn_q80}{Numeric. The 80th percentile of earnings.}
\item{sample_n}{Numeric. The sample size.}
}
}
\source{
U.S. Census Bureau, American Community Survey, Integrated Public Use Microdata Series \url{https://usa.ipums.org/usa/}. Analyzed using the \code{\link{srvyr}} package.
}
\usage{
wages
}
\description{
A dataset containing median individual earnings by various dimensions (education, occupation, etc.) for the US and Maryland, for adults ages 25 and up with positive earnings. The data is calculated from the 2022 American Community Survey (ACS) Public Use Microdata Sample (PUMS) data via the Integrated Public Use Microdata Series (IPUMS).
}
\keyword{datasets}
